added lot of informations, changing a bit the structure of files, started parsing

+++ sources/utils/utils_1.c
@@ -9,7 +9,7 @@
  * buffer pointed by 'buf'.
  */
 
-int ms_strlen(const char *str)
+size_t ms_strlen(const char *str)
 {
 	int 	i;
